// econect-backend/types/scheduleCollect.ts
export interface INewSchedule {
  ad_id: number;
  weight: string;
  point_accumulated: string;
  note: string;
  schedule_date: string;
  schedule_hours: string;
  delivery_type: string;
  operation: string;
  collect_point_id: string;
  type_material_id: string;
  isrecyclable: boolean;
  amount: number;
}